[8.0][mass_mailing_unique] Avoid duplicates.
This new module avoids duplicates in mass mailing lists.

Now you cannot have 2 lists with the same name, and you cannot have the same
email repeated in a list.

This will avoid sending the same mail to a contact repeated times, which will
disturb him and most probably cause you to be blocked as spam.