[8.0][mail_forward] Custom migration
This is a little migration script I need because in Grupo ESOC we had a
slightly different version of this module before publishing to OCA, and
now there's a conflict this solves.

This also serves as an example migration for https://github.com/OCA
/server-tools/pull/369, since it uses that toolkit for doing it. But to
avoid conflicts with others, as you can see, all is inside a
`try...except` that will just allow anyone to run the migration without
hassle.