[fix] mail_digest: move `template_id` to `digest_template_id`
When creating digest records within the context of mail composer
(and possibly other contexts) you'll have a `default_template_id`
key in the context which is going to override our safe default.

This is going to break email generation because the template
will be completely wrong and unpredictable. Lesson learned :)