mass_mailing_resend (#210)
* [ADD] mass_mailing_resend

Resend mass mailings
====================

A frequent need for users of mass mailings is to resend one mailing that has
already been sent in the past to new recipients that haven't received yet that
mail. But the problem is to know which are the applicable ones.

Odoo already includes a method in its mass mailing logic that avoids to resend
the same mail 2 times for one mass mailing, and for v9, there was a trick that
allows to modify the state of a mass mailing from kanban view, covering the
need.

But now on v10 both status bar in form view and dragging between states in
kanban are not allowed.

This module introduces a button to restart a mass mailing to draft state,
allowing you to reevaluate the sending domain or list for performing again
the mailing.

Usage
=====

* Go to *Mass mailing > Mailings > Mass Mailings*.
* Click on one record that is done or create a new one and send it.
* You will see a button called "Resend".
* If you click on it, mass mailing will be set to Draft again.

Known issues / Roadmap
======================

* Add an indicator / filter for knowing resent mailings.
* Include information on the number of new recipients to be sent on the
  resending (through `get_remaining_recipients` method).