[9.0] mail_activity: convert mixin to decorator
The Odoo v9 model mechanism has some issues with loading
AbstractModels as mixins; specifically, the computed fields of the
original model no longer work reliably.

This new approach uses a Python decorator to change the class when
the module is being loaded, so the Odoo model mechanisms only see
a class that does a simple inheritance, and already "comes" with
the new fields/methods related to the activities.

It's essentially the same as copy-pasting the fields and methods
directly into the class.