[IMP] website_mass_mailing_name: Make tour faster
Under some integration environment, the public part of the tour fails to load, I don't know exactly why (or better, I don't know why it was working before, because it was deleting the phantomjs session entirely).

In any case, with this fix, it works there and is faster to test.