[FIX] avoid bounce on 'error' event
Catching this event as a bounce leads quite often to false bounces, like
in the cases the error is due to our own instance SMTP errors. Errors
like a too large message should no bounce a thread's recipients as the
problem isn't in the recipients side but ours.

We leave anyway some true positives aside (like in the case of malformed
email addresses). We should be able to detect and fix those anyway.