[MIG] mail_check_mailbox_size: Migration to 12.0
This module sends an email with a summary of remote IMAP mailbox
folders, with their total space used.

We are going straight to 12.0, but this commit includes the changes for
the 11.0 accorging to the wiki [1]

Also, regenerate README with oca-gen-addon-readme and add myself as
contributor.

Some additional changes to support Python 3:

The imaplib module in Python 3 returns payloads as bytes, which are not
automatically converted to strs. That's a good thing! But we need to
encode manually.

The IMAP4 protocol optionally uses double quoting when mailboxes contain
spaces. This is always allowed. So to fix this, first we use shlex to
perform double-quote-aware splitting (a bit like bash would split
command line arguments) and then we send folder names with quotes to the
IMAP server.

I tested this manually with a couple real IMAP servers.

Co-Authored-By: eLBati <lorenzo.battistini@agilebg.com>

[1] https://github.com/OCA/maintainer-tools/wiki/Migration-to-version-11.0