[FIX] mass_mailing_resend: Incorrect warning text
It's not possible anymore to remove mass mailing stats for tricking the system to
send the same mailing to the already sent recipients, which is, on the other hand,
something not healthy for traceability purposes, so this limitation will continue.

We only change then the warning message for being correct in the statement.

TT25325