[MIG] mail_outbound_static: Improvements to version 10 and 12, it eases the sending of emails from a fixed "from" address to comply with DMARC conformity, use it when your emails can't be sent from Odoo due to invalid "from" address.
The module is very useful as the fix in the Odoo stable version has not being accepted yet.

Compared to v10 and v12, this module adds the feature to allow to setup the "from" address in the server conf file (usually odoo.conf), which is also a valid way for Odoo to setup outgoing email servers.