[IMP] mail_optional_autofollow: add autofollow option on chatter composer
This commit adds an "autofollow" checkbox on the chatter composer.
Because mail_tracking adds the additional recipients (in CCs and such)
as suggested partners, these would be added as followers
when they are replied to.

Note that whatever the autofollow value on the chatter composer,
the full composer will end up with autofollow set to true.
This is because the standard _onOpenFullComposer always pass autofollow
to true in the context, and does not provide any extendable hook.