[FIX] mail_outbound_static: valid email_from
parseaddr does separate the part of an email address and the name but
actually:

1. does not check if the email address is a valid one.
2. does not properly work if the incomming Message['From'] is in a wrong format.

Thanks to this, when we are trying to get the domain of the email_from a really generic exception is raised. "IndexError: list index out of range" only because the email_from is not a valid email address.

To avoid this we check the email_from address is a valid one first.

The test case that brings this change was this one: Message['From'] = "Name, Last Name <valid_email@email.from>.
Since we have the name_from with a comma but without a "" surroding the parseaddr result was ('', 'Name')

Now the proper raise will be shown to the user.