[FIX] mass_mailing_partner: Fix assignation of company name
Defining a res.company on a partner should not set the company_name
on a mailing contact as the mailing contact is not linked to a
res.company.
The company_name should instead be set according to the commercial
partner of the partner being modified if it's parent_id is changed.