[FIX] mail_debrand: default engine as odoo while rendering
On mail_render_mixin._render_template method as we call the
parent method defined in odoo we should use the same
default engine value as mail module. Because `template_src`
can be the content of the template or a reference to a to
a qweb template.

To reproduce this:

* install sms and mail_debrand modules
* go to the partner list and select multiple parteners
* In Action click on *Send SMS Text Message*

Without this fix return content is an empty
string. With the `inline_template` engine
collect body is properly rendered.