Fix mail_composer_cc_bcc conflicting with mass.mailing.
Email from a mass mailing are not being sent. On the related mailing.trace
record the failure type is `Missing email address`.

Uninstalling the module `mail_composer_cc_bcc module` fixes the problem.

The reason this module interfere with sending mass email is that the `mass_mailing`
module uses the mail.compose.message wizard.

The failure type mentionned above happens because the dictionary generated
containing the info for sending the email has no data in the `email_to` key !