[16.0][IMP] mail_debrand: Add remove_href_odoo_special_case
When rendering the template, the remove_href_odoo function will remove the parent of the <a> element which has 'odoo.com' in href.
There is a problem with mail.template "New Portal Signup", the parent element is almost the whole content hence most of the content will be removed.

remove_href_odoo_special_case() is added to handle this special case.t push