[REL] mail_quoted_reply: Migrate JavaScript to 16.0
Rewrote JavaScript to match major changes in 15.0->16.0 transition.
Prior code relied on methods that no longer exist. New code patches the
multiple models in @mail/models to mimic the code present for the
"reply to" MessageAction.

This version also fixes a bug in which the quoted reply button appeared
on notes as well as emails and sometimes duplicated itself repeatedly
after mouseover events.