[16.0][FIX][REF] mail_show_follower: internal notes
This module aims to show the recipients of an email in the email header, but the previous implementation was not suitable for internal notes.

When **logging an internal note** with tagged users and/or partners, those specific recipients should be displayed
as well as the document's followers subscribed to this message subtype.

When **sending an email**, the recipients should be the document's followers.

This is achieved by using the union of `mail.mail`'s `recipient_ids` and `notified_partner_ids`.