[FIX] mail_activity_team: open team activities from systray
** To Reproduce **

1. Create an activity for a team of the current user but with no specific user of for another user in the team.
2. Go to the systray activity widget. Switch to tab Team Activities.
3. Click on `1 Future`.

** Actual outcome **

The activity kanban view opens without any records.

** Expected behavior **

The activity kanban view opens with the team activity that was created in step 1.

** Technical description of the fix **

The systray widget will open the view with a domain that selects the users's activities. Thanks to a context key that is already set, we can intercept the domain and replace it with a domain on the user's teams' activities.

Backported from https://github.com/OCA/mail/pull/38