[17.0][FIX]``mail_activity_team``: do not systematically remove the user from an activity when a team is set
In [this PR](https://github.com/OCA/social/pull/1474), we updated the module to follow the v17 changes in standard.
However it looks like the changes made are too wide: if no user is set, we indeed want to prevent the standard method to set one by default. But if a user is manually added, we want to keep it.
Here we add a context key to determine at the creation of the activity if the user should be kept or erased.
We also take [this issue](https://github.com/OCA/social/issues/1655) into account.