[FIX] mail_gateway: Users without technical access rights can't link a message
Before this change, when a user tried to link a message to a record, the system threw an error because they did not have enough access rights to see the ir.model records.

With this change, the error is no longer raised, and the user can link messages only to the models for which they have write access rights.