[IMP] mail_attach_existing_attachment: Add more information in the attachment display name
When a message contains multiple attachments, including duplicates or files with the same name, it becomes difficult to distinguish them.

This commit improves the display name of attachments by adding the creator (user) and creation date, making it easier to differentiate records.

NOTE: This change impacts the display name at the model level. In the mail scope, this function cannot be easily inherited, which is why the change is applied directly here.
https://github.com/odoo/odoo/blob/7a190bd33d3038828b22575d7aa860eacb503e05/addons/mail/static/src/core/common/composer.js#L505-L544